requirements_clean.txt
accelerate==0.24.1
aiofiles==23.2.1
altair==5.1.2
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
attrs==23.1.0
bitsandbytes==0.41.2.post2
blis==0.7.11
braceexpand==0.1.7
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
cmake==3.27.7
confection==0.1.5
contourpy==1.2.0
cycler==0.12.1
cymem==2.0.8
decorator==4.4.2
decord==0.6.0
docker-pycreds==0.4.0
exceptiongroup==1.1.3
fastapi==0.104.1
ffmpy==0.3.1
filelock==3.13.1
fonttools==4.44.0
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
gradio==3.47.1
gradio_client==0.6.0
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
huggingface-hub==0.17.3
idna==3.4
imageio==2.33.0
imageio-ffmpeg==0.5.1
importlib-resources==6.1.1
iopath==0.1.10
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
lazy_loader==0.3
lit==17.0.4
marisa-trie==1.1.1
MarkupSafe==2.1.3
matplotlib==3.7.0
moviepy==1.0.3
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.2.1
nltk==3.8.1
numpy==1.26.1
omegaconf==2.3.0
opencv-python==4.7.0.72
orjson==3.9.10
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.2
peft==0.2.0
Pillow==10.1.0
portalocker==2.8.2
preshed==3.0.9
proglog==0.1.10
progressbar2==4.2.0
protobuf==4.25.0
psutil==5.9.6
pycparser==2.22
pydantic==2.4.2
pydantic_core==2.10.1
pydub==0.25.1
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-multipart==0.0.6
python-utils==3.8.1
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rpds-py==0.12.0
safetensors==0.4.0
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.3
selenium==4.16.0
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==1.34.0
setproctitle==1.3.3
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.27.0
sympy==1.12
thinc==8.2.3
threadpoolctl==3.2.0
tifffile==2023.9.26
timm==0.6.13
tokenizers==0.14.1
toolz==0.12.0
tqdm==4.66.1
transformers==4.35.0
trio==0.23.1
trio-websocket==0.11.1
triton==2.0.0
typer==0.9.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
uvicorn==0.24.0.post1
visual-genome==1.1.1
wandb==0.16.0
wasabi==1.1.3
weasel==0.3.4
webdataset==0.2.48
websockets==11.0.3
wsproto==1.2.0
zipp==3.17.0
